+++ python/ray/ml/tests/test_data_parallel_trainer.py
@@ -8,7 +8,7 @@ from ray.ml.constants import PREPROCESSOR_KEY
 from ray.ml.train.data_parallel_trainer import DataParallelTrainer
 from ray.ml.preprocessor import Preprocessor
 from ray.tune.tune_config import TuneConfig
-from ray.tune.tuner import Tuner
+from ray.ml.tuner import Tuner
 
 
 @pytest.fixture
