+++ python/ray/ml/tests/test_tuner.py
@@ -14,11 +14,11 @@ from ray.ml.examples.pytorch.torch_linear_example import train_func as linear_tr
 from ray.ml.train.integrations.torch import TorchTrainer
 from ray.ml.train.integrations.xgboost import XGBoostTrainer
 from ray.ml.train import Trainer
+from ray.ml.tuner import Tuner
 from ray.tune import Callback, TuneError
 from ray.tune.cloud import TrialCheckpoint
 from ray.tune.result import DEFAULT_RESULTS_DIR
 from ray.tune.tune_config import TuneConfig
-from ray.tune.tuner import Tuner
 
 
 class TestDatasource(Datasource):
