+++ python/ray/tune/BUILD
@@ -393,14 +393,6 @@ py_test(
 )
 
 py_test(
-    name = "test_tuner",
-    size = "medium",
-    srcs = ["tests/test_tuner.py"],
-    deps = [":tune_lib"],
-    tags = ["team:ml", "exclusive", "tests_dir_T"],
-)
-
-py_test(
     name = "test_utils",
     size = "small",
     srcs = ["tests/test_utils.py"],
